FAQ sectui enhacement - separating CRUD form into its own compoennt for a cleaner codebase
