+++ src/app/modules/faq/components/faq-form/faq-form.component.scss
@@ -1,0 +1,3 @@
+#faq_gridLayout {
+    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+}
