--- src/app/modules/faq/views/index/index.component.scss
@@ -1,3 +1,0 @@
-#faq_gridLayout {
-    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-}
